+++ kernel/src/arch/idt.c
@@ -32,7 +32,7 @@ void idt_install()
 {
     /* Sets the special IDT pointer up, just like in 'gdt.c' */
     idtp.limit = (sizeof (struct idt_entry) * 256) - 1;
-    idtp.base = &idt;
+    idtp.base = (int)&idt;
 
     /* Clear out the entire IDT, initializing it to zeros */
     kmemset(&idt, 0, sizeof(struct idt_entry) * 256);
